refactor(web): split otel-settings.tsx 506->248 LOC (Cluster A2)
WHY: otel-settings.tsx packed two presentational primitives (SettingRow,
Toggle), all form state + Supabase save logic, the env-vars panel, and a static
metrics table into 506 lines - over the 400-LOC ceiling. (The OTEL domain logic
was already factored into lib/otel-config.ts + tested; this is the component
split.) First A2 work in the web package.

Extracted into a co-located otel-settings/ directory:
- parse-headers.ts   - pure header-JSON parser; now directly testable
- useOtelSettings.ts - form state + preset/save/copy handlers + computed preview
- SettingRow.tsx / Toggle.tsx - presentational primitives
- EnvVarsPanel.tsx   - generated env-vars preview + copy
- MetricsReference.tsx - static exported-metrics table
- __tests__/parse-headers.test.ts - 5 tests pinning empty-vs-malformed +
                        the array-rejection fix

Boy-scout fix: the inline parser accepted a JSON array as "headers"
(typeof [] === "object"), which passed the save-gate Object.keys().length check
with bogus data. parseHeaders now rejects arrays + primitives; tested.

The orchestrator (248 LOC) keeps the named OtelSettings export + OtelSettingsProps
at the same path, so the dynamic() import in settings-client.tsx and the
perf-regression bundle-ratchet guard are unchanged. Note: otel-settings.tsx (file)
resolves before otel-settings/ (dir) by module-resolution file-priority - verified
by a clean next build.

Verification: web typecheck + lint clean; full suite 3910 pass (+5 new);
perf-regression dynamic-import guard green; next build exit 0.
